requirements.txt
PyJWT~=2.10.1
wechatpy~=1.8.18
requests~=2.32.4
pycryptodome~=3.23.0
django~=5.2.5
smartdjango

bocfx
beautifulsoup4~=4.13.4
scrapy~=2.13.3
diq~=0.1.1

oba~=0.1.1
PyMySQL~=1.1.1